Update rainy-48px (2).xml
Added a comment to the 'rainy-48px' vector drawable.

The origin and purpose of this vector drawable are currently unknown. The comment serves as a marker and reminder to investigate its source and necessity in the future. If the file is determined to be unnecessary or problematic, it can be safely removed in a subsequent update.